Add sample SourceCred graphs along with an updater
This commit stores a few sample SourceCred graphs (for the major
SourceCred repos themselves), along with an `update.sh` script which
updates the graphs.

The sourcecred/sourcecred repository has been included as a submodule,
to facilitate reproducible usage of the sourcecred CLI at a consistent
checkout hash.

Test plan: The sample graphs in this repository were generated using the
update script.