Initial code for importing cred graphs (#10)
This commit contains inaugural infra code, which imports a SourceCred
graph and adds type annotations. The code itself is quite simple, with a
bit of logic for pulling out the most specific declared node/edge type and
adding it as a property in the networkx graph.

The logic is lightly tested using the sample graphs. We verify that each
graph loads, and for the sourcecred/sourcecred graph, we verify that we
have reasonable counts by node and edge type.

Test plan:

Run the included unit tests via
`python infra/import_graph_test.py`